Latest updates from Astropy affiliated package template

+++ wcsaxes/_astropy_init.py
@@ -111,7 +111,6 @@ def test(package=None, test_path=None, args=None, plugins=None,
         coverage=coverage, open_files=open_files, **kwargs)
 
 if not _ASTROPY_SETUP_:
-
     import os
     from warnings import warn
     from astropy import config
@@ -121,10 +120,19 @@ if not _ASTROPY_SETUP_:
 
     if not os.environ.get('ASTROPY_SKIP_CONFIG_UPDATE', False):
         config_dir = os.path.dirname(__file__)
-        try:
-            config.configuration.update_default_config(__package__, config_dir)
-        except config.configuration.ConfigurationDefaultMissingError as e:
-            wmsg = (e.args[0] + " Cannot install default profile. If you are "
-                    "importing from source, this is expected.")
-            warn(config.configuration.ConfigurationDefaultMissingWarning(wmsg))
-            del e
+        config_template = os.path.join(config_dir, __package__ + ".cfg")
+        if os.path.isfile(config_template):
+            try:
+                config.configuration.update_default_config(
+                    __package__, config_dir, version=__version__)
+            except TypeError as orig_error:
+                try:
+                    config.configuration.update_default_config(
+                        __package__, config_dir)
+                except config.configuration.ConfigurationDefaultMissingError as e:
+                    wmsg = (e.args[0] + " Cannot install default profile. If you are "
+                            "importing from source, this is expected.")
+                    warn(config.configuration.ConfigurationDefaultMissingWarning(wmsg))
+                    del e
+                except:
+                    raise orig_error
